challenges/generator/encoding-002/requirements.txt
Pillow
pydub